Fix incorrect operation ID refresh in log messages
The Fields API was refreshing the log message operation ID
(OID) when the base class did the same immediately. This
makes the Fields API methods not to be part of operation
performance measurements. But the double refresh is not
visible anywhere.

Also if logs are printed from the Fields class and the
refresh is made in Fields, the OID would be different
in the same transaction.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>