Fix Fields API responses for not supported HTTP methods
In case other than GET method was tried to be used with Fields
API, empty response was sent. Now in case keywords, groups or
tags Fields APIs are used with HTTP method that is now allowed,
proper error code 405 Method Not Allowed is sent back in the
response.

The response was take from [1]. Also Falcon sends 405 for HTTP
methods that are not implemented with on_*.

[1] https://stackoverflow.com/a/34324179

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>