Fix test case failure
The test test_cli_create_solution_006 used a fixture that uses a
test helper methods which directly calls Collection() from the
Snippy project. This bypasses all the design principles and only
reason to do this it get things "easily".

Because the code was directly called before the Config() object
was initialized with init(), this caused a call to variable that
is not defined.

The workaround for this is to add 'snippy' figure before calling
another fixture that requires the Collection() usage. Basically
this creates the Snippy twice.

This is not a proper way to do things. But the question how this
should be done without replicating the code to test helper is not
known. Therefore the current situation is accepted "as is".

When running all the cases in row, other cases are creating Snippy().
This seems to work so that the global Config() is persisted somehow
with some values over test cases. Naturally starting the snippy
service in each case overrides previous configuration.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>