Add support to import all default content
It is now possible to import all default content with one
command. Command to import the defaults from all categories is:

  $ snippy import --all --defaults

Import with --all category does not support anything else than
importing the defaults for now. The reason is that the tool is
originally built to process one content category type from one
source file. The yaml and json formatted files already now can
do this but this is not officially supported. The problem is
the text file format which does not support this currently. And
supporting import from a file with --all (or without it) opens
a door to support any content time mixed on one file.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>