Add support for --limit option from CLI interface
The default content limit 20 intended for the REST API server
was too short for the CLI interface. There are more than 20
snippets in the default content and all of them were not
printed which caused confusion.

It is now possible to use --limit option which limits the
number of search result contents printed on terminal. By
default the limit is for the CLI 99 and API 20.

  $ snippy search --sall . --limit 10

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>