Fix parsing of --log-msg-max
Currently this parameter is undocumented option for REST API
server which allows configuring log message maximum length
for -vv (very verbose) option. The parsing of integer value
after the option stored the value accidentally into a list.

This did not cause any problems because the option is read
immediately from sys.argv in order to get the value from
the beginning of the startup for all the logs. Just in case
the Config.log_msg_max is used, the value is now properly
set as integer.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>