Fix Resource() comparison
There were three failures related to comparing Resources.

The resources that were imported from Yaml file were imported so
that the links, tags and groups were set as list. The Resource()
comparison failed because other field was list and other was tuple.
Now the import from JSON or Yaml file always set the groups, tags
and links as a tuple because these must not be changed.

The resource tags and groups that were imported from Yaml or
Json file were not sorted. The default is that these fields are
always stored sorted so this must be the case when importing data
from Yaml or Json file. The links are a special case that they are
not sorted in case of References. The though is that in case of
reference (URL links), the order of links is related. That is, the
first link is more valuable for user than the second link. For
snippets and solutions the links must be sorted. The content
category specific link sort is not completed yet.

The content comparison compared internal data structures when
two resources were tried to compare in __eq__. Now the Resource()
comparison is made which leads to Resource() specific __eq__.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>