Fix internal search keyword management
The search keyword test for database queries was simplified. The
search keywords are sall, stag, sgrp, suuid, sdigest and sdata which
can trigger a search from different fields. There were double checks
or confusing conditions if the search keywords existed. The Config
parser guarantees that these values are either empty tuple or tuple
for sall, stag, sgrp and sdata search keywords. For suuid and sdigest
the value is either None or valid value which can be for example an
empty string.

There is a special case where the value is initially None but this
is taken care by the Config() and the Parser(). The special case is
that the CLI user used one of the search keywords with e.g. --sall
but did not give anything. This must result to search all which is
taken care by setting regexp filter '.' that matches to all content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>