Add support to run test cases in parallel in Python 3
There is now Makefile target 'test-fast' that runs most of the
pytests with all available cores. There are two API server tests
that spawn real process which are skipped with this target.

This works with Python 3. With Python 2 the tool database is
stored as a file because in Python 2 the Sqlite does not support
in-memory database. The filename is same for all tests and thus
it causes problems.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>