Edit search filter variables and comments
Changed the --filter option related internal variable
names to better reflect the usage and to synchronize
the variable names in different modules.

Tried to avoid using term 'filter' in another context
than referring to --filter option in code and test
documentation. Because of this, the usage of this was
removed from test documentation.

The --filter parameter is bit misleading and it behaves the
same as grep. Therefore the current --filter parameter does
not bring any benefit. Because of this, some changes are
needed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>