Change --filter option behaviour
The --filter option used to do the same that could be done
with Linux grep. This was not useful and thus the option
behaviour was changed.

The filter option still consumes a regexp filter. Now the
filter is used so that all content that is matches against
any of the search keys like sall, is filtered. The filter
removes all the content from search results that does not
match to the regexp filter. That is, any field in the
content must match to the regexp otherwise the content
gets removed from the search results.

The previous functionality can be achieved by running Linux
regexp for the CLI command output.

The same option can be now used with the API server. In this
case the behaviour is the same as with the CLI option.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>