Add new content field description (breaking)
New field was added for long description. There already was
brief description that is intended to provide one line
description of content. Now it is possible to add long
description that gives more details about.

This was part of the plan to provide Markdown formatted
content which would contain long description of the
content.

Because of this, there is a new column in the database
file and a new key in the default YAML content files.

Tool upgrade
============

Previously stored content can be imported as before. In
this case, the description field is set to empty string.

It is not possible to use same storage file as before.
Before upgrading the tool, the previously stored content
must be exported to YAML file. After tool is upgraded, the
content can imported from the file.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>