Fix text template parser for first content
There was a bug which caused incorrect content to be read from
text file for the first content in the text source. It is not
sure if this was visible to end user.

Fixed possible case where unknown content could be migrated
to internal Collection data structure. It is not sure if this
was visible to end user.

Refactored text template parser to be object instead of static
class. This tries to abstract the text parser so that it can
return a Collection object which contains all parser resources.
This should make it easier to optimize the text parsing. This
should also allow easier path to implement reading different
content categories from a single file. All this should be
easier because now the whole logical functionality is inside
the ContentParserText() class.

This refactoring prepared to introduce the Markdown parser
in a same manner than the proprietary text template parser.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>