Fix Read the Docs build
There is/was a 'ImportError: cannot import name sqrt' after
recent commit [1]. Root cause is not known. This commit updates
the Sphinx version and updates the Sphinx configuration file.

One suspected reason could be that there was the path setting:

  sys.path.insert(0, os.path.abspath('../../'))

This is now removed from the latest configuration file. There
should be no other major changes.

[1] 2f9278c602e92625447be2e03728b01c240f8170

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>