Change content text template tags
The text templates are tool internal templates and their
functionality is not visible outside. The text template
tags were changed from upper case to lower case. Also tag
names were simplified by removing the repeating 'SNIPPY_'
in front of tags.

This was aimed also to do a bit polish for the templates
before the Markdown specific templates are added.

Because user do not directly manipulate the templates,
this is not a breaking change.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>