Add PyUp service
The PyUp notices outdated Python dependencies and possible
security issues.

The configuration is made so that the PyUp should not create
pull requests automatically. Only the badge is currently
used to check if all dependencies are OK.

The configuration takes automatically the requirements.txt
that contains only the live dependencies. This is intended
behaviour because only the live dependencies are shipped
with a release.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>