Change content template data tag
The solution and snippet template <data> tags was changed.
There is now only one <data> tag in snippet content and in
case of solution, the <data> tags cover the whole template.

This change was done in order to harmonize the logic how to
fill the content template with defined content. With Markdown
syntax, there is a simplified way of adding the data into a
template. This same way is now being taken into use with text
templates.

The solution data tags cover the whole solution template
because the whole data is a solution. For snippet, there is
only one data tag.

The plan is to support:

  1. <data> ... <data>
  2. <data>

formats for the data tags. This can be replaced with one
regexp in all cases. But this generalization required the
solution data tags to cover the whole data (which is also
more logical). At the same time the snippet data tag was
simplified to one tag since it is also more logical.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>