Fix content output indention in terminal output
If there were more than 10 contents printed into terminal,
the output alignment was shifted after tenth item. This was
caused from increased index number. This is now fixed so that
indention is kept no matter how many items are printed.

Changed also snippet content output by adding one empty line
between the header and data in terminal print. This change
harmonizes the output between solutions and references which
use also empty line between header and data.

Refactored the terminal output. The terminal dump was first
tried to be made with templates like with text and Markdown
formats. But this was problematic because of:

  1. Templating required re.sub usage which affected heavily
     to performance. This may have not been relevant but the
     output 'smoothness' was considered important. The change
     caused 4000 snippet terminal output to be delayed from
     0.02 seconds to 0.6 seconds. The current implementation
     manages to print the same set in 0.035 seconds.

  2. Templates from Config would have to be passed to Resource
     class. This is problematic because design allows using
     Resource only from Collection and collection can be imported
     from anywhere. This would have required passing the Config
     templates for example to Editor() which is under Config that
     does not import Config.

     Also the __str__ would require this and passing arguments
     to this method is not possible.

The current way to print the data is much more compact and it
should be more readable than the previous implementation which
used multiple methods to spread the implementation to too many
editor pages.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>