Fix content data update with ASCII escape sequences
If any of the content was updated with editor and the data
container special characters like newlines '\n' or tabs '\t',
the special characters were interpolated.

User defined strings must be processed as raw literal string.
If the user wants to use \n in a snippet, the text must be
stored and presented as is.

The current fix may not be the perfect one. It may be that
there are still cases where same problem occurs. It should
be checked if the user provided content data can be stored
in Resource() data as raw literal string that could be
presented in text without having to escape data string when
outputting them to sting.

On the other hand, this may be a problem only with text
presentation. Internally the data is stored in dictionary
and tuple for data. This likely takes care of the data
format quite nicely and this problem occurs only when data
is printed to terminal or text.

Currently the later case described above is more likely.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>