Change default export file format from text to YAML
When content is exported from command line, the default
file format is now YAML. This is considered more generic
than the text format.

The text format was the first format to be implemented.
When the Markdown is coming the tool specific text format
should be put in the background. The text format can be
exported by defining the filename and format with -f|--file
option as '*.txt' or '*.text'.

Cleaned and simplified the implementation to select the
operation filename.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>