Refactor Dockerfile to reduce image size
The docker container now removes the pip and Python cache
files. This reduced the image size from 58MB to 33MB.

Removing the pip also reduces attack surface so this is
good from security point of view as well.

Based on testing, the startup speed did not suffer and it
was roughly 0.5 seconds with the same search command with
and without Python cache files. The snippy is run two times
to import the defaults after removing all the Python cache
files so only the needed files are compiled in the container.

When running the server, required cache files are generated
into the container when used.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>