Fix Python setup for PyYAML
More flexibility was added to select PyYAML version.

The only dependency is the PyYAML for command line operations.
Python 3.7 requires PyYAML 3.13. The PyYAML 3.12 is few years
old so it should be available easily. Currently the PyYAML
version 4 and above is broken [1].

For some reason the pipenv locked the requirements dependency
file so that the PyYAML is exactly 3.12 even though the Pipfile
defines it as >=3.12. This was left as is in order to avoid
manual changes. Also the requirements.txt is used only in
Docker container which uses latest Python versions so it is
not an issue there. Also the official installations do not
instruct to use the requirements.txt file.

[1] https://github.com/yaml/pyyaml/issues/193

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>