Fix deprecated warnings from tests with Python 3.7
When the test suite was run with the Python 3.7, it produced
few warnings from Falcon and jsonschema packages. Since these
are external dependencies and not related to Snippy code, the
warnings are now ignored.

Pytest does not yet support pyproject.toml so the filtering
rules were set to setup.cfg [1]. When the PEP518 is supported
by Pytest, configuration should be moved to pyproject.toml.

[1] https://github.com/pytest-dev/pytest/issues/1556

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>