Refactor exporting text formatted content to file
The text format export is now following the same design as
the Markdown format export. This uses now text_dump() from
Collection. This changes the functionality so that there is
only one write to file instead of file write for all resources
in collection.

It can be estimated that at least with small or medium size
exports this is faster because it avoids additional disk writes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>