Add support to import content in Markdown format
It is now possible to export and import content in Markdown
format.

Changed the Markdown formatted link presentation. The links
are not inside [`] marks as fixed length code blocks. With
the new approach without the code block, it is possible to
directly click the link in browser.

There is a bug with solution content where the metadata is
cumulated in each export-import iteration. This is because
the whole content including the metadata in mkdn formatted
file is considered as data. This is to be fixed.

Also there is a fault where possible alignment for snippet
data comments gets lost in export-import loop.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>