Edit comments about Markdown formatted solution
Temporary comments were added to fix the regexp that parses
the solution data for Markdown formatted data. The intent is
to support two formats for the solution data. One with the
existing text based code block and full Markdown formatted
solution.

Since it is likely that the Markdown format is so common that
it is preferred over existing text format, two formats are
going to be supported for a bit different behaviour.

Using '# Solution' header in Markdown formatted solution will
allow using full Markdown formatted data without the code block
which would render the actual Markdown format just as a code
block.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>