Change Markdown templates header levels
Previously the Meta header and option Solution headers were
defined as first level headers. It was considered better that
the whole content would fall under one level one header which
is the first line and the content brief. This allows saving
multiple contents into a single file where each content is
separated under level one header.

The Meta and optional Solution header in Markdown content
are now level two headers. It is possible to define these
as level one headers but that is not recommended. First
level headers work but the main development and testing
focus on second level headers.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>