Fix regexp parser for links in Markdown content
The regexp parser for links in Markdown format allowed colon
used in the link like '> [1]: https://'. This is not the way
how the links are written since the colon is not used. The
links are in format '> [1] https://'.

This was fixed so that the parsed does not support colon as
a possible character between index and the link itself when
scanning the document.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>