Fix test cases that use content templates
Previously the content text templates were copied in the
test case helper libraries. This caused faults that were
not detected by test cases when the actual template was
changed but the test helper was not changed. It is much
better to use the real template which forces the test
cases to find faults.

Example test case that did not find fault in described
situation was for example exporting the default template
like in test_cli_create_solution_003.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>