Add support to use description field in templates
It is now possible to add optional content description
from content text templates. Markdown formatted templates
are already supporting the description field. Also the
rest API has already support for the description field.

The solution template has header for description. The
format to add solution description is as in example
below. The description must be after the description
header. The description itself must start with a hash
and it is recommended to use hash for each line of the
description.

---

    # This is a very
    # long description
    # for the solution

---

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>