Fix link parsing from solutions
There was a failure that read incorrect links from text or
Markdown formatted solution template. The intention is to
read only links defined with prefix '> '. There was a failure
in regexp that read links that were prefixed with '>>' or
'  '.

Only user defined links with prefix ' > ' must be parsed
to Resource().

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>