Add support to auto align comments in snippets
When snippet content is printed to terminal, the comments are
automatically aligned when the content is printed to terminal.
There is no need for the user to try to align the comments.

It is possible to add comments only for portion of the commends.
The alignment is made based on only the commands that have a
comment. This tries to avoid aligning comments too far right
and to keep the line length in control.

The hash between command and comment is colored in order to
visually separate these two items.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>