Fix Collection data class usage
The Collection data class is not being used as data class.
The supported ways to iterate and list resources stored in
collection can be seen from tests_ut_collection test case.

Collection resource meta data was removed because it did not
provide any additional functionality with benefits. Initially
the resource specific meta data was added to support special
case where the user did not request digest field which was
removed. This has been refactored long time ago and this is
not current behaviour which made the resource meta useless.

The collection meta is still in the code but it is going
to be removed because that is not needed anymore since the
len(collection) returns the amount of resources in collection.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>