Fix Collection data class usage
Removed the Collection.data because it was not supposed to
be used externally. The data class offers the keys, values
and resources methods to access stored resources.

Added more test and fixed merge method for collection. The
merge() for collection is not currently used but it was
decided to be left as of now. The update operation supports
only update of one resource of the time. The update calls
directly to Resource().merge and it bypasses the merge in
Collection(). It is being considered that the update could
support Collection so the merge was left as of now.

The collection.total is actually the total amount of
resources without filter. This attribute was added because
of a special case to convey the total of search hits from
database to REST API. This is not logical and it should be
changed to something else.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>