Fix text wrapping of long description in Markdown
The content description in Markdown was not wrapped when
the Markdown was presented by some online resources like
GitHub. The intention is to have hard breaks in description
that force the line breaks instead of leaving this to the
end application.

The problem with the additional spaces in the Markdown at
the end of line is that these are easily lost when editors
remove them automatically. Currently the only clean way
to force the line break for metadata that works with all
online resources seen so far is the two trailing spaces
trick.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>