Refactor test cases Content compared away
There is currently the Content.compared which forces the
default content to be in the same format than in database.
The Reference default content was made so that there are
intentional errors in the default content. For example the
tags are not sorted in the default content. This was created
to test error scenarios with the default content.

This was a mistake that creates problems.

The main problem is that most of the tests are not testing
error scenarios but this forces always the default content
to be manipulated for content comparison. This also hides
the error testing behind a wrapper. The proper way should
be that the error scenario is clearly defined in the test
cases itself.

This also part of the target to refactor the tests to
clarify the 'testing API' and method hidden behind the
Content() testing class.

- Refactor tests to use Content() compare_mkdn, compare_text, compare_yaml, compare_json to compare exported data in various formats.
- Refactor tests to use Content() compare_data to compare content in database.
- Refactor and clarify the need to use Content() ordered for REST API test to very JSON results.
- Refactor tests to use the new mocked_file which allows defining the return format for the mock.
- Refactor the failure cases from default content and move the failures explicitly to tests.
- Refactor result text compare to own Pytest plugin to have own tests and evolution.
- Refactor common items to Helper class. The Helper class must not have dependencies in order to be imported anywhere.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>