Refactor database content comparison in test cases
There is now new method intended to test created or updated
content in database. The new method compares all but the key
attribute between reference and database collections. The
previous method used text based comparison and it tested only
the fields that are visible in the text template.

The new method is also roughly 16 times faster than the previous
method when comparison single test case with single snippet
content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>