Fix description field indenting in Markdown content
The long description field was incorrectly indented in case
of Markdown formatted content. The two spaces meant to be at
the end of all but the very last line of description were
incorrectly set. The spaces were added at the beginning of
each line.

The two spaces are needed with some Markdown formatters. Most
of the online formatters work with out this. The problem is
at least with GitHub showing the content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>