Fix solution data trimming
The solution data was trimmed for every line from right. It
is considered that in case of solution, user wants to keep
data field without any modifications. The only modification
to solution data is that there will be only one newline at
the end of the solution data. The intention is to remove
unnecessary empty lines with just a linefeed characters at
the end of solution.

Refactored the text and markdown content data parsing so
the data sanitization is clearly made when the data is
returned from the parser. There were also few missing places
where the sanitization was not made for text and markdown
content.

Added log prints to print the content if the parsing fails
for text and markdown content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>