Edit TODO list and remove DONE items
Remove unnecessary entries from the list. Test case refactoring
is ongoing and only one line is needed.

The storage assert was updated to contain also the dictionary
based comparison. The problem with only comparing Collections
was that the collection fields are e.g. trimmed and sorted.
This meant that all faults are not visible by only comparing
Collections.

The new Content assert methods solve also the old problem
with comparing error title field. Previously all error title
fields were ignore and not compared. Now all but the JSON
schema validation failures are compared. The problem with
the JSON schema validation error is that it is different
between Python 2 and Python 3 due to Unicode strings. Also
the JSON schema validation _may_ contain random order
dictionary or list in a string which make the comparison a
bit more complicated and thus it is skipped as of now.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>