Fix CLI Unicode character decoding for Python 2
The Unicode decoding seemed to actually work when testing
from command line. But one of the Unicode tests failed for
Python 2 after refactoring. This should have worked because
the CLI input parameter was converted to Unicode even though
the type was 'str'. Author is not totally confident of this
explanation because the 'str' should (?) fail with Unicode
character on Python 2. Maybe the CLI input test with western
setting was somehow hiding the problem.

Now the CLI input is converted directly to Unicode with both
python major versions. The conversion is needed only with
Python 2 but the same 'to_unicode' can handle both cases.

This is also now much more clearer because the type indicates
directly on main cli.py module that the handling takes care
of Python 2.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>