Fix content update when parsing user input fails
There was an exception thrown for the user if content was
updated with editor and when the parser failed to read the
user input. This applied to all parsing failures with text,
Markdown, JSON or YAML format. The relevant scenarios for
this fault are the text and Markdown based content updates.

If the content parsing fails, the content is not updated in
the database.

There is now proper exit cause printed for the user.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>