Edit Python module versions for development
Updated Pytest and Pytest-xdist to latest versions.

This change impacted also requirements.txt. This file has
the dependencies for server application without development
dependencies. The updated six in the requirements.txt was
coming from updated Pytest version which is not application
dependency. But because the pipenv is used to to generate
the requirements.txt file, the six version was updated there.

The changelog of six has new features and it [1] does not
list any bug fixes.

This update also fixes the Pyup warning in Snippy project
README badge.

[1] https://github.com/benjaminp/six/blob/d927b9e27617abca8dbf4d66cc9265ebbde261d6/CHANGES

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>