Change UTC offset format from +0000 to +00:00 (breaking)
All timestamps including logs, default content and code was
changed to use UTC offset format with colon delimiter. This
was made in order to support CockroachDB/psycopg2 which do
not support UTC offset format '+0000'.

The existing content must be exported to YAML formatted file
where every timestamp must be manually changed from '+0000'
to '+00:00'. The tool would parse currently existing content
correctly. But since new format was taken into use, the old
format is not officially supported.

  $ snippy export --all --defaults
  $ # Manually change time timezone format.
  $ # Manually delete the existing content under snippy/data/storage.
  $ snippy import --all --defaults

Python parses UTC timezone offset with colon only from Python
3.7 onwards [1, 2, 3i, 4]. With Python3.6 and earlier, you can
use datetime to parse format '+0000' only. This means that when
using Python3.6 and ealier, you must replace the '+00:00' with
'+0000'.

  if Const.PYTHON37:
      try:
          datetime.datetime.strptime(timestamp, '%Y-%m-%dT%H:%M:%S.%f%z')
      except ValueError:
          return False
  elif not Const.PYTHON2:
      try:
          timestamp = timestamp.replace('+00:00', '+0000')
          datetime.datetime.strptime(timestamp, '%Y-%m-%dT%H:%M:%S.%f%z')
      except ValueError:
          return False
  else:
      timestamp = timestamp[:-6]  # Remove last '+00:00'.
      try:
          datetime.datetime.strptime(timestamp, '%Y-%m-%dT%H:%M:%S.%f')
      except ValueError:
          return False

The JSON formatted logs have all timestamps in UTC time format
and text logs in local time. This tries to improve readability
while debugging and have common timestamp when running the server
in live.

The timestamp could have used just Z which is the same as '+0000'.
For the human (author's mind) readability it was chosen to use
the '+0000' instead of just adding the Z.

[1] https://stackoverflow.com/a/48539157
[2] https://github.com/python/cpython/commit/32318930da70ff03320ec50813b843e7db6fbc2e
[3] https://stackoverflow.com/a/49784038
[4] https://bugs.python.org/issue31800

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>