Refactor test helpers to always use Collection()
Some of the test case helpers were accessing Resource()
directly. All the code should always use the Collection().

This allows removing load_dict() method from the Resource
which was used only by the test helpers. Now the resource
can be added only by migrating or merging a resource.

This should clarify the implementation and help for example
in digest computation optimization. In this case it should
be possible to reduce failure branches to compute and check
digest if the resource migrate and merge always guarantee
correct resource digest.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>