Fix parsing snippet with hash comments
If a snippet was added with a comment mark in the first row, the
snippet was read. It is not sure if it is a good idea to support
snippets like:

  # List directory
  ls -al

The intention is to support snippets like below where each line
contains valid comment like:

  ls -al # List directory

For now, the snippets are parsed with the comment mark.

There was also a failure in parsing regexp. The regexp used DOTALL
flag combined with ^. The caret ^ was incorrectly used to search
line after a newline. But since the DOTALL was used, the caret does
not match immediately after the newline but the beginning of the
whole string.

Changed also one cause indicating internal error bad request. The
cause for parsing failure is not necessary a failure in the tool
and thus the bad request was considered more appropriate.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>