Refactor content validation before storing it
Previously the content to be stored in the database was
validated in the database layer. This was not a good idea
when bringing more options for the backend database. Also
it is considered better to validate user content outside
of the database layer and leave only database specific
checks like integrity errors into the database layer.

This is part of refactoring to bring support for new
database options on top of Sqlite.

This also changed slightly the handling of cause code
settings in error scenarios. The changes were made in
order to make the new implementation possible.

This commit will fail test since one fix was wanted to be
left for separate commit.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>