Fix raising exception
If same exception is raised again, it must be done just with
raise. If the exception name is added at the end of raise
command, all the parameters from the original exception are
lost.

This was misunderstanding that is now corrected. This had an
affect to new implementation where the content validation was
moved out from database layer. This did not have impact in
the previous implementation.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>