Refactor database insert to use batch insert
The database insert was changed to insert multiple rows
in one row at the time. This helps significantly when
importing content with PC's with limited disk throughput.
This does not help that much with other operations and
with PC's that already have fast disks. There may be
bigger benefit also with modern PC's when importing
hundreds or thousands of rows.

This also aims to refactor the database layer to support
PostgreSQL and CockroachDB as a server backends.

If any of the rows in bulk insert fail due to database
integrity error when violating already existing UUID or
digest, each row is tried to be inserted one by one. If
any of the rows is inserted successfully, the insert
operation is results 201 Created. Each failing content
will produce error log and cause which is visible in the
logs.

The REST API server returns only the created content so
user is able to know if all content was insert or not.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>